make commands queue a property of connection object instad of global var
